fix: update UI QR Code
