+++ apps/mobile/src/app/screens/settings/MyQRCode/index.tsx
@@ -1,16 +1,21 @@
 import { size, useTheme } from '@mezon/mobile-ui';
 import { selectAllAccount, useWallet } from '@mezon/store-mobile';
 import { CURRENCY, createImgproxyUrl, formatBalanceToString } from '@mezon/utils';
-import { useCallback, useEffect, useMemo, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useTranslation } from 'react-i18next';
-import { Image, Share, Text, TouchableOpacity, View } from 'react-native';
+import { Text, TouchableOpacity, View } from 'react-native';
 import { Grid } from 'react-native-animated-spinkit';
 import FastImage from 'react-native-fast-image';
+import Share from 'react-native-share';
+import Toast from 'react-native-toast-message';
 import { useSelector } from 'react-redux';
+import RNFetchBlob from 'rn-fetch-blob';
 import RNQRGenerator from 'rn-qr-generator';
 import MezonIconCDN from '../../../componentUI/MezonIconCDN';
 import { IconCDN } from '../../../constants/icon_cdn';
 import { useImage } from '../../../hooks/useImage';
+import type { CustomQRInviteRef } from '../../home/homedrawer/components/InviteToChannel/CustomQRInvite';
+import CustomQRInvite from '../../home/homedrawer/components/InviteToChannel/CustomQRInvite';
 import { style } from './styles';
 
 type TabType = 'profile' | 'transfer';
@@ -22,28 +27,32 @@ interface QRCode {
 
 export const MyQRCode = () => {
 	const { themeValue } = useTheme();
-	const styles = style(themeValue) as any;
-	const { t } = useTranslation(['profile']);
+	const styles = style(themeValue);
+	const { t } = useTranslation(['profile', 'inviteToChannel']);
 	const userProfile = useSelector(selectAllAccount);
 	const { saveMediaToCameraRoll } = useImage();
 	const [activeTab, setActiveTab] = useState<TabType>('profile');
 	const [isGenerating, setIsGenerating] = useState<boolean>(true);
 	const [qrCode, setQrCode] = useState<QRCode>({ profile: '', transfer: '' });
 	const { walletDetail } = useWallet();
+	const profileQRRef = useRef<CustomQRInviteRef>(null);
+	const transferQRRef = useRef<CustomQRInviteRef>(null);
 
 	const tokenInWallet = useMemo(() => {
-		return walletDetail?.balance || 0;
+		return walletDetail?.balance || '0';
 	}, [walletDetail?.balance]);
 
 	const profilePayload = useMemo(() => {
 		try {
-			const payload = {
-				id: userProfile?.user?.id,
-				avatar: userProfile?.user?.avatar_url,
-				name: userProfile?.user?.display_name
-			};
-
-			const encodedPayload = btoa(encodeURIComponent(JSON.stringify(payload)));
+			const encodedPayload = btoa(
+				encodeURIComponent(
+					JSON.stringify({
+						id: userProfile?.user?.id,
+						avatar: userProfile?.user?.avatar_url,
+						name: userProfile?.user?.display_name
+					})
+				)
+			);
 			return `${process.env.NX_CHAT_APP_REDIRECT_URI}/chat/${userProfile?.user?.username}?data=${encodedPayload}`;
 		} catch (error) {
 			console.error('Error QR Profile Payload', error);
@@ -52,22 +61,24 @@ export const MyQRCode = () => {
 	}, [userProfile?.user?.id, userProfile?.user?.avatar_url, userProfile?.user?.username, userProfile?.user?.display_name]);
 
 	const transferPayload = useMemo(() => {
-		const data = {
+		return JSON.stringify({
 			receiver_name: userProfile?.user?.username,
 			receiver_id: userProfile?.user?.id
-		};
-		return JSON.stringify(data);
+		});
 	}, [userProfile?.user?.id, userProfile?.user?.username]);
+
+	const isShowActionButton = useMemo(() => {
+		return !isGenerating && qrCode?.[activeTab] && activeTab === 'profile';
+	}, [activeTab, isGenerating, qrCode]);
 
 	const generateQRCode = async (type: TabType) => {
 		try {
-			const payload = type === 'profile' ? profilePayload : transferPayload;
 			setIsGenerating(true);
 			const res = await RNQRGenerator.generate({
-				value: payload,
-				height: Math.ceil(activeTab === 'profile' ? size.s_400 : size.s_220),
-				width: Math.ceil(activeTab === 'profile' ? size.s_400 : size.s_220),
-				correctionLevel: 'L'
+				value: type === 'profile' ? profilePayload : transferPayload,
+				height: Math.ceil(type === 'profile' ? size.s_400 : size.s_220),
+				width: Math.ceil(type === 'profile' ? size.s_400 : size.s_220),
+				correctionLevel: 'Q'
 			});
 			setQrCode((pre) => ({
 				...pre,
@@ -79,29 +90,47 @@ export const MyQRCode = () => {
 		}
 	};
 
-	const handleDownloadQRCode = useCallback(async () => {
+	const handleDownloadQRProfile = useCallback(async () => {
 		try {
-			const qrCodeUri = qrCode?.['profile'];
-			if (!qrCodeUri) return;
-			const filePath = qrCodeUri.startsWith('file://') ? qrCodeUri : `file://${qrCodeUri}`;
-			await saveMediaToCameraRoll(filePath, 'image', true, false);
-		} catch (e) {
-			console.error('QR Code download error:', e);
-		}
-	}, [qrCode, saveMediaToCameraRoll]);
+			const qrRef = activeTab === 'profile' ? profileQRRef : transferQRRef;
+			const brandedUri = await qrRef.current?.capture();
+			if (!brandedUri) return;
 
-	const handleShareQRCode = useCallback(async () => {
-		try {
-			const qrCodeUri = qrCode?.['profile'];
-			if (!qrCodeUri) return;
-			await Share.share({
-				url: qrCodeUri,
-				message: profilePayload || ''
-			});
-		} catch (e) {
-			console.error('QR Code share error:', e);
+			const filePath = brandedUri.startsWith('file://') ? brandedUri : `file://${brandedUri}`;
+			await saveMediaToCameraRoll(filePath, 'image', true, false);
+		} catch (error) {
+			console.error('QR Code download error:', error);
 		}
-	}, [profilePayload, qrCode]);
+	}, [activeTab, saveMediaToCameraRoll]);
+
+	const handleShareQRProfile = useCallback(async () => {
+		try {
+			const brandedUri = await profileQRRef.current?.capture();
+			if (!brandedUri) return;
+
+			const baseDir = `${RNFetchBlob.fs.dirs.CacheDir}/mezon_qr`;
+			const exists = await RNFetchBlob.fs.exists(baseDir);
+			if (!exists) await RNFetchBlob.fs.mkdir(baseDir);
+
+			const shareFilePath = `${baseDir}/qr_share_profile.png`;
+			const fileExists = await RNFetchBlob.fs.exists(shareFilePath);
+			if (!fileExists) await RNFetchBlob.fs.cp(brandedUri.replace('file://', ''), shareFilePath);
+
+			await Share.open({
+				url: `file://${shareFilePath}`,
+				type: 'image/png',
+				title: `QR_Profile`,
+				message: 'Scan QR code to chat with me on Mezon',
+				failOnCancel: false
+			});
+		} catch (error) {
+			Toast.show({
+				type: 'error',
+				text1: t('inviteToChannel:qrModal.shareError')
+			});
+			console.error('Error sharing QR code:', error);
+		}
+	}, []);
 
 	useEffect(() => {
 		if (!qrCode?.[activeTab]) {
@@ -120,9 +149,9 @@ export const MyQRCode = () => {
 
 	const userInfo = useMemo(
 		() => ({
-			avatarUrl: userProfile?.user?.avatar_url,
-			username: userProfile?.user?.username,
-			displayName: userProfile?.user?.display_name
+			avatarUrl: userProfile?.user?.avatar_url || '',
+			username: userProfile?.user?.username || '',
+			displayName: userProfile?.user?.display_name || userProfile?.user?.username || ''
 		}),
 		[userProfile?.user?.avatar_url, userProfile?.user?.username, userProfile?.user?.display_name]
 	);
@@ -130,8 +159,8 @@ export const MyQRCode = () => {
 	return (
 		<View style={styles.container}>
 			<View style={styles.tabContainer}>
-				{renderTabButton('profile', t('qr_profile', 'QR Profile'))}
-				{renderTabButton('transfer', t('qr_transfer', 'QR Transfer'))}
+				{renderTabButton('profile', t('qr_profile'))}
+				{renderTabButton('transfer', t('qr_transfer'))}
 			</View>
 
 			<View style={[styles.card]}>
@@ -145,16 +174,16 @@ export const MyQRCode = () => {
 						/>
 					) : (
 						<View style={styles.defaultAvatar}>
-							<Text style={styles.textAvatar}>{userInfo.username?.charAt?.(0)?.toUpperCase()}</Text>
+							<Text style={styles.textAvatar}>{userInfo.username?.charAt?.(0)?.toUpperCase() || ''}</Text>
 						</View>
 					)}
 
 					<View>
-						<Text style={styles.nameProfile}>{userInfo.username || userInfo.displayName}</Text>
+						<Text style={styles.nameProfile}>{userInfo.displayName}</Text>
 						<Text style={styles.tokenProfile}>
 							{activeTab === 'profile'
 								? `${t('shareWithOthers')}`
-								: `${t('token')} ${formatBalanceToString(tokenInWallet || '0')?.toString()} ${CURRENCY.SYMBOL}`}
+								: `${t('token')} ${formatBalanceToString(tokenInWallet)} ${CURRENCY.SYMBOL}`}
 						</Text>
 					</View>
 				</View>
@@ -162,29 +191,28 @@ export const MyQRCode = () => {
 				<View style={styles.qrContainer}>
 					{isGenerating ? (
 						<Grid color={themeValue.text} size={size.s_50} />
-					) : (
-						<View style={styles.qrWrapper}>
-							<Image source={{ uri: qrCode?.[activeTab] }} style={styles.imageQR} />
-						</View>
-					)}
+					) : qrCode?.[activeTab] ? (
+						<CustomQRInvite
+							ref={activeTab === 'profile' ? profileQRRef : transferQRRef}
+							qrCodeUri={qrCode[activeTab]}
+							clanLogo={userInfo.avatarUrl}
+							clanName={userInfo.username}
+						/>
+					) : null}
 				</View>
 
-				{!isGenerating && qrCode?.[activeTab] && activeTab === 'profile' && (
+				{isShowActionButton && (
 					<View style={styles.actionsRow}>
-						<TouchableOpacity style={styles.actionButton} onPress={handleDownloadQRCode}>
+						<TouchableOpacity style={styles.actionButton} onPress={handleDownloadQRProfile}>
 							<MezonIconCDN icon={IconCDN.downloadIcon} color={themeValue.text} />
 						</TouchableOpacity>
-						<TouchableOpacity style={styles.actionButton} onPress={handleShareQRCode}>
+						<TouchableOpacity style={styles.actionButton} onPress={handleShareQRProfile}>
 							<MezonIconCDN icon={IconCDN.shareIcon} color={themeValue.text} />
 						</TouchableOpacity>
 					</View>
 				)}
 				<View style={styles.descriptionContainer}>
-					<Text style={styles.descriptionText}>
-						{activeTab === 'profile'
-							? t('qr_profile_description', 'Scan this QR code to chat with me or view my profile')
-							: t('qr_transfer_description', 'Scan this QR code to transfer funds')}
-					</Text>
+					<Text style={styles.descriptionText}>{activeTab === 'profile' ? t('qr_profile_description') : t('qr_transfer_description')}</Text>
 				</View>
 			</View>
 		</View>
